Feature additons for Elmo and Result path
Elmo path was hardcoded. Changed to accept arguments. Saved as hyperparameters (not so efficient. But this works)
While evaluation, the result was stored in temporary path and deleted. But test results are important for analysis. Only while using the model for testing, the result path is passed as parameter and saved